+++ src/PPU/CMakeLists.txt
@@ -1,3 +1,3 @@
-ADD_LIBRARY(PPU STATIC PPU.cpp)
+ADD_LIBRARY(PPU STATIC ppu.cpp)
 
 TARGET_LINK_LIBRARIES(PPU GBComponent)
